move NullOrdering to the Dialect package where its usages are
